Fix e2e test case
