+++ cypress/e2e/billingentities.cy.ts
@@ -1,8 +1,6 @@
 import { createUser } from '../fixtures/user';
 import { billingEntityNxt, billingEntityVshn, setBillingEntities } from '../fixtures/billingentities';
 import { BillingEntityPermissions } from '../../src/app/types/billing-entity';
-import { createClusterRoleBinding } from '../fixtures/clusterrole-binding';
-import { ClusterRoleBinding, ClusterRoleBindingPermissions } from '../../src/app/types/clusterrole-binding';
 
 describe('Test billing entity list', () => {
   beforeEach(() => {
@@ -81,7 +79,7 @@ describe('no permissions', () => {
     cy.get('h1').should('contain.text', 'Welcome to the APPUiO Cloud Portal');
   });
 
-  it('no member view permission', () => {
+  it('no member edit permission', () => {
     cy.visit('/billingentities/be-2345/members');
     cy.get('h1').should('contain.text', 'Welcome to the APPUiO Cloud Portal');
   });
@@ -120,247 +118,5 @@ describe('Test billing entity details', () => {
     cy.get('.flex-wrap > .text-900').eq(3).should('contain.text', '📃📋🏤 🏙️🇨🇭');
     cy.get('.flex-wrap > .text-900').eq(4).should('contain.text', 'mig 📧');
     cy.get('.flex-wrap > .text-900').eq(5).should('contain.text', '🇩🇪');
-  });
-});
-
-describe('billing entity edit members', () => {
-  beforeEach(() => {
-    cy.setupAuth();
-    window.localStorage.setItem('hideFirstTimeLoginDialog', 'true');
-    cy.disableCookieBanner();
-  });
-  beforeEach(() => {
-    // needed for initial getUser request
-    cy.intercept('GET', 'appuio-api/apis/appuio.io/v1/users/mig', {
-      body: createUser({ username: 'mig', defaultOrganizationRef: 'nxt' }),
-    });
-    cy.setPermission(
-      { verb: 'list', ...BillingEntityPermissions },
-      { verb: 'get', ...ClusterRoleBindingPermissions, name: 'billingentities-be-2345-viewer' },
-      { verb: 'get', ...ClusterRoleBindingPermissions, name: 'billingentities-be-2345-admin' },
-      { verb: 'update', ...ClusterRoleBindingPermissions, name: 'billingentities-be-2345-viewer' },
-      { verb: 'update', ...ClusterRoleBindingPermissions, name: 'billingentities-be-2345-admin' }
-    );
-  });
-
-  it('add member', () => {
-    cy.intercept('GET', 'appuio-api/apis/billing.appuio.io/v1/billingentities/be-2345', {
-      body: billingEntityNxt,
-    });
-    cy.intercept(
-      'GET',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-viewer',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-viewer', users: ['appuio#mig'] }),
-      }
-    );
-    cy.intercept(
-      'GET',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-admin',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-admin', users: ['appuio#mig'] }),
-      }
-    );
-    cy.intercept(
-      'PUT',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-viewer',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-viewer', users: ['appuio#mig', 'appuio#crc'] }),
-      }
-    ).as('updateViewer');
-    cy.intercept(
-      'PUT',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-admin',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-admin', users: ['appuio#mig', 'appuio#crc'] }),
-      }
-    ).as('updateAdmin');
-
-    cy.visit('/billingentities/be-2345/members');
-    cy.get('.text-3xl').should('contain.text', 'be-2345 Members');
-    cy.get('[data-cy="name-input-1"]').type('crc');
-    cy.get('p-multiselect').eq(1).click().contains('billingentities-be-2345-admin').click();
-    cy.get('button[type=submit]').click();
-    cy.wait('@updateViewer');
-    cy.wait('@updateAdmin');
-    cy.get('@updateViewer')
-      .its('request.body')
-      .then((body: ClusterRoleBinding) => {
-        expect(body.subjects).to.have.length(2);
-        const subject = body.subjects && body.subjects[1];
-        expect(subject && subject.name).to.eq('appuio#crc');
-      });
-    cy.get('@updateAdmin')
-      .its('request.body')
-      .then((body: ClusterRoleBinding) => {
-        expect(body.subjects).to.have.length(2);
-        const subject = body.subjects && body.subjects[1];
-        expect(subject && subject.name).to.eq('appuio#crc');
-      });
-  });
-
-  it('delete member', () => {
-    cy.intercept('GET', 'appuio-api/apis/billing.appuio.io/v1/billingentities/be-2345', {
-      body: billingEntityNxt,
-    });
-    cy.intercept(
-      'GET',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-viewer',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-viewer', users: ['appuio#mig', 'appuio#crc'] }),
-      }
-    );
-    cy.intercept(
-      'GET',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-admin',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-admin', users: ['appuio#mig', 'appuio#crc'] }),
-      }
-    );
-    cy.intercept(
-      'PUT',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-viewer',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-viewer', users: ['appuio#mig'] }),
-      }
-    ).as('updateViewer');
-    cy.intercept(
-      'PUT',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-admin',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-admin', users: ['appuio#mig'] }),
-      }
-    ).as('updateAdmin');
-
-    cy.visit('/billingentities/be-2345/members');
-    cy.get('.text-3xl').should('contain.text', 'be-2345 Members');
-    cy.get('button[title="Remove"]').eq(1).click();
-    cy.get('button[type=submit]').click();
-    cy.wait('@updateViewer');
-    cy.wait('@updateAdmin');
-    cy.get('@updateViewer')
-      .its('request.body')
-      .then((body: ClusterRoleBinding) => {
-        expect(body.subjects).to.have.length(1);
-        const subject = body.subjects && body.subjects[0];
-        expect(subject && subject.name).to.eq('appuio#mig');
-      });
-    cy.get('@updateAdmin')
-      .its('request.body')
-      .then((body: ClusterRoleBinding) => {
-        expect(body.subjects).to.have.length(1);
-        const subject = body.subjects && body.subjects[0];
-        expect(subject && subject.name).to.eq('appuio#mig');
-      });
-  });
-
-  it('delete member shows warning if removing own user', () => {
-    cy.intercept('GET', 'appuio-api/apis/billing.appuio.io/v1/billingentities/be-2345', {
-      body: billingEntityNxt,
-    });
-    cy.intercept(
-      'GET',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-viewer',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-viewer', users: ['appuio#mig', 'appuio#crc'] }),
-      }
-    );
-    cy.intercept(
-      'GET',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-admin',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-admin', users: ['appuio#mig', 'appuio#crc'] }),
-      }
-    );
-
-    cy.visit('/billingentities/be-2345/members');
-    cy.get('.text-3xl').should('contain.text', 'be-2345 Members');
-    cy.get('button[title="Remove"]').eq(0).click();
-    cy.get('p-message').contains('You are about to remove yourself as admin!');
-  });
-
-  it('unassign role', () => {
-    cy.intercept('GET', 'appuio-api/apis/billing.appuio.io/v1/billingentities/be-2345', {
-      body: billingEntityNxt,
-    });
-    cy.intercept(
-      'GET',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-viewer',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-viewer', users: ['appuio#mig', 'appuio#crc'] }),
-      }
-    );
-    cy.intercept(
-      'GET',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-admin',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-admin', users: ['appuio#mig', 'appuio#crc'] }),
-      }
-    );
-    cy.intercept(
-      'PUT',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-viewer',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-viewer', users: ['appuio#mig', 'appuio#crc'] }),
-      }
-    ).as('updateViewer');
-    cy.intercept(
-      'PUT',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-admin',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-admin', users: ['appuio#mig'] }),
-      }
-    ).as('updateAdmin');
-
-    cy.visit('/billingentities/be-2345/members');
-    cy.get('.text-3xl').should('contain.text', 'be-2345 Members');
-    const multiSelect = cy.get('p-multiselect').eq(1);
-    multiSelect.click();
-    multiSelect.get('p-multiselectitem').contains('billingentities-be-2345-admin').click();
-    cy.get('button[type=submit]').click();
-    cy.wait('@updateViewer');
-    cy.wait('@updateAdmin');
-    cy.get('@updateViewer')
-      .its('request.body')
-      .then((body: ClusterRoleBinding) => {
-        expect(body.subjects).to.have.length(2);
-        const mig = body.subjects && body.subjects[0];
-        const crc = body.subjects && body.subjects[1];
-        expect(mig && mig.name).to.eq('appuio#mig');
-        expect(crc && crc.name).to.eq('appuio#crc');
-      });
-    cy.get('@updateAdmin')
-      .its('request.body')
-      .then((body: ClusterRoleBinding) => {
-        expect(body.subjects).to.have.length(1);
-        const subject = body.subjects && body.subjects[0];
-        expect(subject && subject.name).to.eq('appuio#mig');
-      });
-  });
-
-  it('unassign role shows warning if removing own user', () => {
-    cy.intercept('GET', 'appuio-api/apis/billing.appuio.io/v1/billingentities/be-2345', {
-      body: billingEntityNxt,
-    });
-    cy.intercept(
-      'GET',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-viewer',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-viewer', users: ['appuio#mig', 'appuio#crc'] }),
-      }
-    );
-    cy.intercept(
-      'GET',
-      'appuio-api/apis/rbac.authorization.k8s.io/v1/clusterrolebindings/billingentities-be-2345-admin',
-      {
-        body: createClusterRoleBinding({ name: 'billingentities-be-2345-admin', users: ['appuio#mig', 'appuio#crc'] }),
-      }
-    );
-    cy.visit('/billingentities/be-2345/members');
-    cy.get('.text-3xl').should('contain.text', 'be-2345 Members');
-    const multiSelect = cy.get('p-multiselect').eq(0);
-    multiSelect.click();
-    multiSelect.get('p-multiselectitem').contains('billingentities-be-2345-admin').click();
-    cy.get('p-message').contains('You are about to remove yourself as admin!');
   });
 });
